A few tweaks to docs
