+++ docs/components/header.css
@@ -10,6 +10,7 @@
   transition: background-color 0.3s ease, color 0.3s ease;
   z-index: 1000;
   border-bottom: 1px solid #e9e9e9;
+  background-color: #fff;
 }
 
 .navContent {
@@ -90,58 +91,6 @@
   background-color: transparent;
 }
 
-.mobileMenuContainer {
-  position: fixed;
-  left: 0;
-  right: 0;
-  height: 100%;
-  background-color: #2f419b;
-  z-index: 1001;
-  display: flex;
-  flex-direction: column;
-}
-
-.mobileMenu {
-  position: fixed;
-  top: 60px;
-  color: #fff;
-  width: 100%;
-}
-
-.mobileMenuDivider {
-  width: calc(100% - 80px);
-  margin: 0 auto;
-  height: 1px;
-  background-color: rgba(223, 223, 223, 0.2);
-}
-
-.mobileMenuItem {
-  padding: 20px 40px;
-  color: #fff;
-  font-size: 14px;
-  text-transform: uppercase;
-  font-weight: 600;
-}
-
-.mobileMenuActions {
-  margin-top: auto;
-  width: 100%;
-}
-
-.mobileMenuAction {
-  width: calc(100% - 80px);
-  margin: 0 auto;
-  margin-bottom: 20px;
-}
-.mobileMenuAction > a,
-.mobileMenuAction > a > button {
-  width: 100%;
-}
-
-.showMobile {
-  display: none;
-}
-
 .docs_tag {
   color: #5e73e5;
   font-size: 20px;
@@ -151,21 +100,16 @@
   margin-top: 18px;
 }
 
-@media (max-width: 1250px) {
-  .showMobile {
-    display: block;
-  }
-  .hideMobile {
-    display: none;
-  }
-  .nav {
-    border-bottom: none;
-  }
-
+@media (max-width: 800px) {
   .navInverted > .navContent > span > a > .logoImage {
     padding: headerPadding;
   }
   .logoImageWhite {
     padding: headerPadding;
   }
+}
+
+.githubLogo {
+  filter: invert(100%);
+  height: 25px;
 }
