+++ docs/components/navigation.css
@@ -8,7 +8,6 @@
 
 .menu {
   background-color: transparent;
-  font-size: 14px;
   color: #2f2d44;
 }
