+++ docs/global.css
@@ -48,6 +48,7 @@
   color: colorsTextMain;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  font-size: 14px;
 }
 
 :global body input {
@@ -107,18 +108,17 @@
   color: #2f2d44;
 }
 
-:global body p, :global body code {
+:global body p,
+:global body code {
   font-style: normal;
   font-weight: normal;
   line-height: 30px;
-  font-size: 15px;
   color: #373e42;
 }
 :global body li {
   font-style: normal;
   font-weight: normal;
   line-height: 30px;
-  font-size: 15px;
   color: #373e42;
 }
