+++ fe/BUILD
@@ -11,24 +11,3 @@ css_typings(
     name = "css_typings",
     srcs = [":css"],
 )
-
-load("@npm_bazel_typescript//:index.bzl", "ts_library")
-
-ts_library(
-    name = "fe",
-    srcs = glob([
-        "**/*.tsx",
-        "**/*.ts",
-    ]),
-    data = [
-        ":css",
-    ],
-    module_name = "@dataform/fe",
-    deps = [
-        ":css_typings",
-        "@npm//@blueprintjs/core",
-        "@npm//@types/node",
-        "@npm//@types/react",
-        "@npm//react",
-    ],
-)
